requirements.txt
mediapipe==0.10.9
firebase_admin
PyYaml
ffmpeg
gunicorn
google-cloud-pubsub